Created prototype for vis 2 and added list to home
The first prototype of the adjacency matrix works and can be found on the vis2 page. I also got enthusiastic and designed a list for the visualizations to show on the homepage. Also made a bunch of small changes to the CSS, including a fix that forces the footer to stay at the bottom of the page