Added numbers and new normalization curve
Changed the fixed-header view to also display numbers. This part took a long time since Django doesn't allow Python functions like zip and enumerate in templates, and it took some time to find a workaround. Also changed the normalization factor, which was plain wrong. Alpha values are between 0-1, not 0-255. The linear normalization is fixed and a new logarithmic normalization formula has been added. Also added the setup for click interactions.