Basic implementation of AM injection
I attempted to inject the AM into the node-link diagram's iframe to put it below the controls and next to the node-link diagram, but setting up the update functions now has to be done within the iframe's context. This'll take a lot more unnecessary work, so I consider it a failed experiment. Committed for future reference.