Applied control styling and fixed sorting
The vis1 controls are now styled within the Dash app instead of using injection, so it's applied more smoothly and globally. Sorting was also incorrect (1 < 100 < 2), so that's been fixed as well using natural sorting. The sorting steps take longer because of this change, and since the algorithm is based on insertion sort and changing directions corresponds to sorting an inversely sorted array, the worst input, yielding an n^2 runtime, the limit has been changed to 100 items